requirements.txt
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.1
scipy>=1.13.0
matplotlib>=3.9.0
yfinance>=0.2.54
requests>=2.32.0
ibapi>=9.81.1.post1
